requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
tldextract>=5.1.0
markdown2>=2.4.0
trafilatura>=7.0.0
pydantic>=2.0.0
jinja2>=3.1.0
google-search-results>=2.4.0
python-dotenv>=1.0.0
weasyprint>=60.0.0
streamlit>=1.28.0
openai>=1.0.0
matplotlib
PyMuPDF
pytesseract
Pillow>=3.7.0
